requirements.txt
ultralytics
torch
torchvision
matplotlib
pillow
opencv-python
norfair
pydantic_settings
jupyter
google-cloud-storage
google-cloud-pubsub